all bugs fixed?
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/ibm@231130 138bc75d-0d04-0410-961f-82ee72b054a4

+++ kelvin-1/gcc/cfgloopmanip.h
@@ -64,6 +64,6 @@ struct loop * loop_version (struct loop *, void *,
 #ifdef KELVIN_PATCH
 extern bool in_loop_p (basic_block, struct loop *);
 extern void zero_loop_frequencies (struct loop *);
-extern void increment_loop_frequencies (struct loop *, basic_block, int);
+extern void increment_loop_frequencies (struct loop *, basic_block, float);
 #endif
 #endif /* GCC_CFGLOOPMANIP_H */
